backend/requirements_api.txt
fastapi>=0.110
sqlalchemy>=2.0
PyMySQL>=1.1
python-multipart>=0.0.9
PyJWT>=2.8
argon2-cffi>=23.1
pydantic[email]>=2.6
python-dotenv>=1.0
opencv-python>=4.9
numpy>=1.26
Pillow>=10.0
torch>=2.2
torchvision>=0.17
ultralytics>=8.2
